backend/order-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
stripe
fpdf
PyJWT
